chore: improve test and tweaks
